+++ src/waterfall.h
@@ -17,7 +17,7 @@ extern float waterfall_auto_min;
 extern float waterfall_auto_max;
 
 lv_obj_t * waterfall_init(lv_obj_t * parent);
-void waterfall_data(float *data_buf, uint16_t size);
+void waterfall_data(float *data_buf, size_t size);
 void waterfall_set_height(lv_coord_t h);
 void waterfall_clear();
 void waterfall_band_changed();
@@ -26,3 +26,4 @@ void waterfall_change_max(int16_t d);
 void waterfall_change_min(int16_t d);
 void waterfall_change_freq(int16_t df);
 void waterfall_update_band(uint64_t f);
+
